clean: Drop unused import
