+++ src/stm32loader/bootloader.py
@@ -23,7 +23,6 @@ import enum
 import math
 import operator
 import struct
-import sys
 import time
 from functools import lru_cache, reduce
